refactor :: oauth, auth count class util로 이동
