+++ src/main/kotlin/com/msg/gauth/domain/auth/repository/MinuteSignInCountRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.auth.repository
 
-import com.msg.gauth.domain.auth.MinuteSignInCount
+import com.msg.gauth.global.util.count.auth.MinuteSignInCount
 import org.springframework.data.repository.CrudRepository
 
 interface MinuteSignInCountRepository : CrudRepository<MinuteSignInCount, String> {
