+++ src/main/kotlin/com/msg/gauth/domain/auth/repository/RefreshTokenRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.auth.repository
 
-import com.msg.gauth.domain.auth.RefreshToken
+import com.msg.gauth.global.util.count.auth.RefreshToken
 import org.springframework.data.repository.CrudRepository
 
 interface RefreshTokenRepository: CrudRepository<RefreshToken, Long>{
