+++ src/main/kotlin/com/msg/gauth/domain/auth/repository/SecondSignInCountRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.auth.repository
 
-import com.msg.gauth.domain.auth.SecondSignInCount
+import com.msg.gauth.global.util.count.auth.SecondSignInCount
 import org.springframework.data.repository.CrudRepository
 
 interface SecondSignInCountRepository : CrudRepository<SecondSignInCount, String> {
