+++ src/main/kotlin/com/msg/gauth/domain/auth/repository/TempSignInBanRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.auth.repository
 
-import com.msg.gauth.domain.auth.TempSignInBan
+import com.msg.gauth.global.util.count.auth.TempSignInBan
 import org.springframework.data.repository.CrudRepository
 
 interface TempSignInBanRepository : CrudRepository<TempSignInBan, String> {
