+++ src/main/kotlin/com/msg/gauth/domain/oauth/repository/MinuteOAuthSignInCountRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.oauth.repository
 
-import com.msg.gauth.domain.oauth.MinuteOAuthSignInCount
+import com.msg.gauth.global.util.count.oauth.MinuteOAuthSignInCount
 import org.springframework.data.repository.CrudRepository
 
 interface MinuteOAuthSignInCountRepository : CrudRepository<MinuteOAuthSignInCount, String> {
