+++ src/main/kotlin/com/msg/gauth/domain/oauth/repository/OauthCodeRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.oauth.repository
 
-import com.msg.gauth.domain.oauth.OauthCode
+import com.msg.gauth.global.util.count.oauth.OauthCode
 import org.springframework.data.repository.CrudRepository
 
 interface OauthCodeRepository: CrudRepository<OauthCode, String> {
