+++ src/main/kotlin/com/msg/gauth/domain/oauth/repository/OauthRefreshTokenRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.oauth.repository
 
-import com.msg.gauth.domain.oauth.OauthRefreshToken
+import com.msg.gauth.global.util.count.oauth.OauthRefreshToken
 import org.springframework.data.repository.CrudRepository
 
 interface OauthRefreshTokenRepository: CrudRepository<OauthRefreshToken, Long> {
