+++ src/main/kotlin/com/msg/gauth/domain/oauth/repository/SecondOAuthSignInCountRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.oauth.repository
 
-import com.msg.gauth.domain.oauth.SecondOAuthSignInCount
+import com.msg.gauth.global.util.count.oauth.SecondOAuthSignInCount
 import org.springframework.data.repository.CrudRepository
 
 interface SecondOAuthSignInCountRepository : CrudRepository<SecondOAuthSignInCount, String> {
