+++ src/main/kotlin/com/msg/gauth/domain/oauth/repository/TempOAuthSignInBanRepository.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.oauth.repository
 
-import com.msg.gauth.domain.oauth.TempOAuthSignInBan
+import com.msg.gauth.global.util.count.oauth.TempOAuthSignInBan
 import org.springframework.data.repository.CrudRepository
 
 interface TempOAuthSignInBanRepository : CrudRepository<TempOAuthSignInBan, String> {
