+++ src/main/kotlin/com/msg/gauth/global/util/count/auth/RefreshToken.kt
@@ -1,4 +1,4 @@
-package com.msg.gauth.domain.auth
+package com.msg.gauth.global.util.count.auth
 
 import com.msg.gauth.global.security.jwt.JwtTokenProvider
 import org.springframework.data.annotation.Id
