+++ src/main/kotlin/com/msg/gauth/global/util/count/auth/util/TooManyRequestValidUtil.kt
@@ -1,11 +1,11 @@
-package com.msg.gauth.domain.oauth.util
+package com.msg.gauth.global.util.count.auth.util
 
-import com.msg.gauth.domain.oauth.MinuteOAuthSignInCount
-import com.msg.gauth.domain.oauth.SecondOAuthSignInCount
-import com.msg.gauth.domain.oauth.exception.OAuthSignInMinuteOverException
-import com.msg.gauth.domain.oauth.exception.OAuthSignInSecondOverException
-import com.msg.gauth.domain.oauth.repository.MinuteOAuthSignInCountRepository
-import com.msg.gauth.domain.oauth.repository.SecondOAuthSignInCountRepository
+import com.msg.gauth.global.util.count.auth.MinuteSignInCount
+import com.msg.gauth.global.util.count.auth.SecondSignInCount
+import com.msg.gauth.domain.auth.exception.SignInMinuteCountOverException
+import com.msg.gauth.domain.auth.exception.SignInSecondCountOverException
+import com.msg.gauth.domain.auth.repository.MinuteSignInCountRepository
+import com.msg.gauth.domain.auth.repository.SecondSignInCountRepository
 import com.msg.gauth.domain.user.User
 import com.msg.gauth.domain.user.enums.UserState
 import com.msg.gauth.domain.user.exception.UserNotFoundException
@@ -14,44 +14,43 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 
 @Component
-class TooManyOAuthRequestValidUtil(
-    private val minuteSignInCountRepository: MinuteOAuthSignInCountRepository,
-    private val secondSignInCountRepository: SecondOAuthSignInCountRepository,
+class TooManyRequestValidUtil(
+    private val minuteSignInCountRepository: MinuteSignInCountRepository,
+    private val secondSignInCountRepository: SecondSignInCountRepository,
     private val userRepository: UserRepository
 ) {
     fun validRequest(email: String){
         val secondSignInCount = (secondSignInCountRepository.findByIdOrNull(email)
-            ?: secondSignInCountRepository.save(SecondOAuthSignInCount(email)))
+            ?: secondSignInCountRepository.save(SecondSignInCount(email)))
 
         if (secondSignInCount.count >= 20) {
             val user = userRepository.findByEmail(email)
                 ?: throw UserNotFoundException()
 
-            userOAuthBan(user)
+            userSignInBan(user)
 
-            throw OAuthSignInSecondOverException()
+            throw SignInSecondCountOverException()
         }
-
         secondSignInCount.addCount()
+
         secondSignInCountRepository.save(secondSignInCount)
 
         val minuteSignInCount = (minuteSignInCountRepository.findByIdOrNull(email)
-            ?: minuteSignInCountRepository.save(MinuteOAuthSignInCount(email)))
+            ?: minuteSignInCountRepository.save(MinuteSignInCount(email)))
 
         if (minuteSignInCount.count >= 10) {
             val user = userRepository.findByEmail(email)
                 ?: throw UserNotFoundException()
 
-            userOAuthBan(user)
-
-            throw OAuthSignInMinuteOverException()
+            userSignInBan(user)
+            throw SignInMinuteCountOverException()
         }
 
         minuteSignInCount.addCount()
         minuteSignInCountRepository.save(minuteSignInCount)
     }
 
-    fun userOAuthBan(user: User){
+    fun userSignInBan(user: User){
         userRepository.save(
             User(
                 id = user.id,
@@ -63,7 +62,7 @@ class TooManyOAuthRequestValidUtil(
                 classNum = user.classNum,
                 num = user.num,
                 roles = user.roles,
-                state = UserState.OAUTH_BAN,
+                state = UserState.SIGN_IN_BAN,
                 profileUrl = user.profileUrl,
                 wrongPasswordCount = user.wrongPasswordCount,
                 oauthWrongPasswordCount = user.oauthWrongPasswordCount
