+++ src/main/kotlin/com/msg/gauth/global/util/count/oauth/OauthCode.kt
@@ -1,4 +1,4 @@
-package com.msg.gauth.domain.oauth
+package com.msg.gauth.global.util.count.oauth
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.redis.core.RedisHash
